+++ src/migrations/m181018_081114_encrypt.php
@@ -1,0 +1,41 @@
+<?php
+
+/**
+ * @copyright  Copyright (c) Flipbox Digital Limited
+ * @license    https://flipboxfactory.com/software/patron/license
+ * @link       https://www.flipboxfactory.com/software/patron/
+ */
+
+namespace flipbox\patron\migrations;
+
+use craft\db\Migration;
+use flipbox\patron\Patron;
+use flipbox\patron\records\Provider;
+
+class m181018_081114_encrypt extends Migration
+{
+    /**
+     * @inheritdoc
+     */
+    public function safeUp()
+    {
+        $this->alterColumn(
+            Provider::tableName(),
+            'clientSecret',
+            $this->char(Provider::CLIENT_SECRET_LENGTH)
+        );
+
+        // Encrypt those that are not
+        if (Patron::getInstance()->getSettings()->encryptStorageData === true) {
+            Patron::getInstance()->manageProviders()->changeEncryption(true);
+        }
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function safeDown()
+    {
+        return true;
+    }
+}